chore(roll): roll to 1.31.0-beta-1676906983000

+++ playwright/_impl/_api_structures.py
@@ -185,7 +185,7 @@ class ExpectedTextValue(TypedDict, total=False):
 class FrameExpectOptions(TypedDict, total=False):
     expressionArg: Any
     expectedText: Optional[List[ExpectedTextValue]]
-    expectedNumber: Optional[int]
+    expectedNumber: Optional[float]
     expectedValue: Optional[Any]
     useInnerText: Optional[bool]
     isNot: bool
